+++ deltachat-jsonrpc/src/api/mod.rs
@@ -39,7 +39,7 @@ use types::webxdc::WebxdcMessageInfo;
 
 use self::types::{
     chat::{BasicChat, MuteDuration},
-    message::MessageViewtype,
+    message::{MessageNotificationData, MessageViewtype},
 };
 
 #[derive(Clone, Debug)]
@@ -641,6 +641,16 @@ impl CommandApi {
             );
         }
         Ok(messages)
+    }
+
+    /// Fetch info desktop needs for creating a notification for a message
+    async fn message_get_notification_info(
+        &self,
+        account_id: u32,
+        message_id: u32,
+    ) -> Result<MessageNotificationData> {
+        let ctx = self.get_context(account_id).await?;
+        MessageNotificationData::from_msg_id(&ctx, MsgId::new(message_id)).await
     }
 
     /// Delete messages. The messages are deleted on the current device and