+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### API-Changes
 - jsonrpc: add `mailingListAddress` property to `FullChat` #3607
+- jsonrpc: add `MessageNotificationData` & `messageGetNotificationInfo()`
 
 ### Changes
 - truncate incoming messages by lines instead of just length #3480